requirements.txt
# Bill Split Backend Requirements
Django>=5.1.6,<5.2
djangorestframework>=3.14.0,<3.15
django-cors-headers>=4.3.1,<4.4
python-dotenv>=1.0.1,<1.1
openai>=1.12.0,<1.13
pytesseract>=0.3.10,<0.4
Pillow>=10.2.0,<10.2
google-generativeai>=0.3.2,<0.8
plotly>=5.18.0,<5.19
gunicorn>=21.2.0,<21.3
tqdm>=4.66.0,<4.67
httpx>=0.27.0,<0.28
rsa>=4.9,<4.10
google-cloud-vision>=3.7.1

